+++ tb_plugin/torch_tb_profiler/profiler/multiprocessing.py
@@ -1,0 +1,4 @@
+import os
+
+def get_start_method():
+    return os.getenv('TORCH_PROFILER_START_METHOD')
